% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{activity_codes}
\alias{activity_codes}
\title{Activity Codes}
\format{
A data frame code 3 columns
}
\usage{
activity_codes
}
\description{
ATOC activity codes matche to GTFS pickup and drop off values
}
\keyword{datasets}
