% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_naptan.R
\name{as_data_table_naptan_stop_area}
\alias{as_data_table_naptan_stop_area}
\title{as data table naptan stop area}
\usage{
as_data_table_naptan_stop_area(doc)
}
\arguments{
\item{doc}{xml document node}
}
\value{
data table of stop areas
}
\description{
Unpacks selected naptan XML doc elements into data.table
}
