% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_naptan.R
\name{as_data_table_naptan_stop_point}
\alias{as_data_table_naptan_stop_point}
\title{as data table naptan stop point}
\usage{
as_data_table_naptan_stop_point(doc, stopTypes = c("RLY"))
}
\arguments{
\item{doc}{xml document node}

\item{stopTypes}{list of stop types to restrict processing to (defaults to railway station)}
}
\value{
data table of stop points
}
\description{
Unpacks selected naptan XML doc elements into data.table
}
\details{
RLY stop types include TIPLOC & CRS fields. The quality of the geographic location is better than from BPLAN
}
