% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{atco_areas}
\alias{atco_areas}
\title{ATCO Admin Areas}
\format{
A SF data frame
}
\usage{
atco_areas
}
\description{
Association of Transport Co-ordinating Officers Code
}
\details{
Boundaries of the ATCO Admin Areas. Note there are 4 national areas
represented by a box around the UK.
}
\keyword{datasets}
