% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{atoc_agency}
\alias{atoc_agency}
\title{Agency.txt for ATOC data}
\format{
A data frame
}
\usage{
atoc_agency
}
\description{
The ATOC data does not included sufficient information to build agency.txt
So this data is provided in the package.
}
\keyword{datasets}
