% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atoc_shapes.R
\name{ATOC_shapes}
\alias{ATOC_shapes}
\title{Make shapes.txt for train GTFS}
\usage{
ATOC_shapes(gtfs)
}
\arguments{
\item{gtfs}{a gtfs object}
}
\value{
Returns a gtfs object
}
\description{
Make shapes.txt for train GTFS
}
\details{
Uses a internal map of the UK rail network to build shapes.txt
}
