% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transxchange_import_functions.R
\name{clean_sequence2}
\alias{clean_sequence2}
\title{Clean NA from sequence}
\usage{
clean_sequence2(x, y, displace = FALSE)
}
\arguments{
\item{x}{sequency of numbers}

\item{y}{sequence of ids showing when to start a new sequency}

\item{displace}{if TRUE start at 2 rather than 2}
}
\description{
Clean NA from sequence
}
