% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_test_files.R
\name{dl_example_file}
\alias{dl_example_file}
\title{Download Test Files}
\usage{
dl_example_file(path, type = "atoc")
}
\arguments{
\item{path}{path to folder}

\item{type}{If "atoc" then downloads ATOC CIF, else TransXchange}
}
\description{
Download small examples of ATOC CIF and TransXchange files. These are mainly
meant for internal testing. But can also be used to check the package is
correctly installed.
}
