% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cal.R
\name{get_bank_holidays}
\alias{get_bank_holidays}
\title{Get Bank Holiday Calendar
TODO: Get scottosh holidays
Downloads and formats the bank holiday calendar for use with TransXchange data.}
\usage{
get_bank_holidays(
  url = "https://www.gov.uk/bank-holidays/england-and-wales.ics"
)
}
\arguments{
\item{url}{url to ics file}
}
\description{
Get Bank Holiday Calendar
TODO: Get scottosh holidays
Downloads and formats the bank holiday calendar for use with TransXchange data.
}
