% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs_cleaning.R
\name{gtfs_clean}
\alias{gtfs_clean}
\title{Clean simple errors from GTFS files}
\usage{
gtfs_clean(gtfs)
}
\arguments{
\item{gtfs}{gtfs list}
}
\description{
Clean simple errors from GTFS files
}
\details{
Task done:

1. Remove stops with no location information
2. Remove stops that are never used
3. Replace missing agency names with "MISSINGAGENCY"
}
