% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs_cleaning.R
\name{gtfs_fast_stops}
\alias{gtfs_fast_stops}
\title{Find fast stops}
\usage{
gtfs_fast_stops(gtfs, maxspeed = 83)
}
\arguments{
\item{gtfs}{list of gtfs tables}

\item{maxspeed}{the maximum allowed speed in metres per second default 83 m/s
(about 185 mph the max speed of trains on HS1 line)}
}
\description{
A varient of gtfs_fast_trips that can detect stops that may be in the wrong location
}
\details{
The function looks a straight line distance between each stop and
  detects the fastest segment of the journey. A common cause of errors is
  that a stop is in the wrong location so a bus can appear to teleport
  across the country in seconds.
}
