% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs_validate.R
\name{gtfs_force_valid}
\alias{gtfs_force_valid}
\title{Force a GTFS to be valid bu removing problems}
\usage{
gtfs_force_valid(gtfs)
}
\arguments{
\item{gtfs}{gtfs object}
}
\description{
Force a GTFS to be valid bu removing problems
}
