% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs_cleaning.R, R/gtfs_split.R
\name{gtfs_split}
\alias{gtfs_split}
\title{Split a GTFS object based on trip_ids}
\usage{
gtfs_split(gtfs, n_split = 2)

gtfs_split(gtfs, n_split = 2)
}
\arguments{
\item{gtfs}{gtfs list}

\item{trip_ids}{a vector of trips ids}

\item{gtfs}{a gtfs object}
}
\description{
Split a GTFS object based on trip_ids

Split a gtfs file
}
