% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs_cleaning.R
\name{gtfs_split_ids}
\alias{gtfs_split_ids}
\title{Split a GTFS object based on trip_ids}
\usage{
gtfs_split_ids(gtfs, trip_ids)
}
\arguments{
\item{gtfs}{gtfs list}

\item{trip_ids}{a vector of trips ids}
}
\value{
Returns a named list of two gtfs objects. The `true` list contains
  trips that matched `trip_ids` the `false` list contains all other trips.
}
\description{
Split a GTFS object based on trip_ids
}
