% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs_to_sf.R
\name{gtfs_stops_sf}
\alias{gtfs_stops_sf}
\title{Make stops into an SF object}
\usage{
gtfs_stops_sf(gtfs)
}
\arguments{
\item{gtfs}{a gtfs object}
}
\value{
sf data frame of points
}
\description{
Make stops.txt into a sf data frame for plotting
}
