% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs_subset.R
\name{gtfs_subset}
\alias{gtfs_subset}
\title{Split a gtfs file by location}
\usage{
gtfs_subset(gtfs, bounds)
}
\arguments{
\item{gtfs}{a gtfs object}

\item{bounds}{an sf object}
}
\description{
Split a gtfs file by location
}
