% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs_to_sf.R
\name{gtfs_trips_sf}
\alias{gtfs_trips_sf}
\title{Make trips into an SF object}
\usage{
gtfs_trips_sf(gtfs)
}
\arguments{
\item{gtfs}{a gtfs object}
}
\value{
sf data frame of lines
}
\description{
Make trips.txt into a sf data frame for plotting
}
