% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs_validate.R
\name{gtfs_validate_external}
\alias{gtfs_validate_external}
\title{Validate GTFS with Google Validator}
\usage{
gtfs_validate_external(path_gtfs, path_validator)
}
\arguments{
\item{path_gtfs}{path to gtfs.zip file}

\item{path_validator}{path to google validator}
}
\description{
Validate GTFS with Google Validator
}
