% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs_write.R
\name{gtfs_write}
\alias{gtfs_write}
\title{Write GTFS}
\usage{
gtfs_write(
  gtfs,
  folder = getwd(),
  name = "gtfs",
  stripComma = FALSE,
  stripTab = FALSE,
  stripNewline = FALSE,
  quote = TRUE
)
}
\arguments{
\item{gtfs}{named list of data.frames}

\item{folder}{folder to save the gtfs file to}

\item{name}{the name of the zip file without the .zip extension, default "gtfs"}

\item{stripComma}{logical, should commas be stripped from text, default = FALSE}

\item{stripTab}{logical, should tabs be stripped from text, default = FALSE}

\item{stripNewline}{logical, should newlines be stripped from text, default = FALSE}

\item{quote}{logical, should strings be quoted, default = TRUE, passed to data.table::fwrite}
}
\description{
Takes a list of data frames representing the GTFS format and saves them as GTFS
Zip file.
}
