% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{historic_bank_holidays}
\alias{historic_bank_holidays}
\title{Historic Bank Holidays}
\format{
A data frame with 4 columns
}
\usage{
historic_bank_holidays
}
\description{
Bank holidays from 2001 to 2018 in the UK. Note Wales has the same holidays
as England.
}
\keyword{datasets}
