% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atoc_import.R
\name{importALF}
\alias{importALF}
\title{Import the .alf file}
\usage{
importALF(file)
}
\arguments{
\item{file}{Path to .alf file}
}
\description{
Import the .alf file
}
\details{
Imports the .alf file and returns data.frame
}
