% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atoc_import.R
\name{importFLF}
\alias{importFLF}
\title{Import the .flf file}
\usage{
importFLF(file)
}
\arguments{
\item{file}{Path to .flf file}
}
\description{
Import the .flf file
}
\details{
Imports the .flf file and returns data.frame
}
