% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atoc_import.R
\name{importMCA}
\alias{importMCA}
\title{Import the .mca file}
\usage{
importMCA(
  file,
  silent = TRUE,
  ncores = 1,
  full_import = FALSE,
  working_timetable = FALSE
)
}
\arguments{
\item{file}{Path to .mca file}

\item{silent}{logical, should messages be displayed}

\item{ncores}{number of cores to use when parallel processing}

\item{full_import}{import all data, default FALSE}

\item{working_timetable}{use rail industry scheduling times instead of public times}
}
\description{
Import the .mca file
}
\details{
Imports the .mca file and returns data.frame
}
