% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atoc_import.R
\name{importMSN}
\alias{importMSN}
\title{Import the .msm file}
\usage{
importMSN(file, silent = TRUE)
}
\arguments{
\item{file}{Path to .msn file}

\item{silent}{logical, should messages be displayed}
}
\value{
named list of data frames: station, timetable, comment, and alias
}
\description{
Import the .msm file
}
\details{
Imports the .msn file and returns data.frame
}
