% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atoc_import.R
\name{importTSI}
\alias{importTSI}
\title{Import the .tsi file}
\usage{
importTSI(file)
}
\arguments{
\item{file}{Path to .tsi file}
}
\description{
Import the .tsi file
}
\details{
Imports the .tsi file and returns data.frame
}
