% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{naptan_missing}
\alias{naptan_missing}
\title{Bus Stop Locations}
\format{
A data frame of 5 columns
}
\usage{
naptan_missing
}
\description{
A database of bus stops that are missing from the NAPTAN but are known to
have been used. For some reason the offical NAPTAN file is missing a small
number of bus stops. This file contains a selection of bus stops that have
appears in TransXchange files, but are missing in the NAPTAN. The have been
assempled from a rnage of sources and may be of varying quality.
}
\keyword{datasets}
