% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{naptan_replace}
\alias{naptan_replace}
\title{Bus Stop Locations wrong in the NaPTAN}
\format{
A data frame of 5 columns
}
\usage{
naptan_replace
}
\description{
A database of bus stops that are wrong in the NAPTAN. These have been
corrected with a mix of manual and automatic techniques. Contributions are
welcome of improved locations.
}
\keyword{datasets}
