% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netex_fares_read.R
\name{netex_read_failures}
\alias{netex_read_failures}
\title{Failures from the last multi-file NeTEx read}
\usage{
netex_read_failures(netex_list)
}
\arguments{
\item{netex_list}{output of [netex_read_fares_multiple()].}
}
\value{
a data.table with columns `file` and `error`.
}
\description{
Extracts the table of files that could not be parsed from an object returned
by [netex_read_fares_multiple()].
}
