% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nptdr.R
\name{nptdr2gtfs}
\alias{nptdr2gtfs}
\title{Import Nataional Public Transport Data Repository (NPTDR) data as GTFS}
\usage{
nptdr2gtfs(
  path = "D:/OneDrive - University of Leeds/Data/UK2GTFS/NPTDR/October-2006.zip",
  silent = FALSE,
  n_files = NULL,
  enhance_stops = TRUE,
  naptan = get_naptan(),
  year = nptdr_archive_year(path)
)
}
\arguments{
\item{path}{Path to zipped folder for NPTDR data}

\item{silent}{Logical, should messages be returned}

\item{n_files}{debug option numerical vector for files to be passed e.g. 1:10}

\item{enhance_stops}{Logical, if TRUE will download current NaPTAN to add in any missing stops}

\item{naptan}{Naptan Locations from get_naptan()}

\item{year}{Year of the NPTDR archive, used to apply standard_mode_overrides() rules}
}
\value{
A gtfs object: a named list of data frames representing the tables
  of a GTFS file
}
\description{
Import Nataional Public Transport Data Repository (NPTDR) data as GTFS
}
\details{
NPTDR is a UK-wide dataset of timetables, 
and is available from data.gov.uk. The data is distributed as a zip file 
containing a number of CIF files, one for each local authority area. 
The function imports the CIF files and returns a GTFS object.
}
