% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atoc_nr.R
\name{nr2gtfs}
\alias{nr2gtfs}
\title{ATOC to GTFS (Network Rail Version)}
\usage{
nr2gtfs(
  path_in,
  silent = TRUE,
  ncores = 1,
  locations = "tiplocs",
  agency = "atoc_agency",
  shapes = FALSE,
  working_timetable = FALSE,
  public_only = TRUE
)
}
\arguments{
\item{path_in}{Character, path to Network Rail ATOC file e.g."C:/input/toc-full.CIF.gz"}

\item{silent}{Logical, should progress messages be surpressed (default TRUE)}

\item{ncores}{Numeric, When parallel processing how many cores to use
(default 1)}

\item{locations}{where to get tiploc locations (see details)}

\item{agency}{where to get agency.txt (see details)}

\item{shapes}{Logical, should shapes.txt be generated (default FALSE)}

\item{working_timetable}{Logical, should WTT times be used instead of public times (default FALSE)}

\item{public_only}{Logical, only return calls/services that are for public passenger pickup/set down (default TRUE)}
}
\value{
A gtfs list
}
\description{
Convert ATOC CIF files from Network Rail to GTFS
}
\details{
Locations

The .msn file contains the physical locations of stations and other TIPLOC
codes (e.g. junctions). However, the quality of the locations is often poor
only accurate to about 1km and occasionally very wrong. Therefore, the
UK2GTFS package contains an internal dataset of the TIPLOC locations with
better location accuracy, which are used by default.

However you can also specify `locations = "file"` to use the TIPLOC locations
in the ATOC data or provide an SF data frame of your own.

Agency

The ATOC files do not contain the necessary information to build the
agency.txt file. Therefore this data is provided with the package. You can
also pass your own data frame of agency information.
}
\seealso{
Other main: 
\code{\link{atoc2gtfs}()}
}
\concept{main}
