% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RailDataPortal.R
\name{nrdp_fares}
\alias{nrdp_fares}
\title{Download Fares  from National Rail Data Portal}
\usage{
nrdp_fares(
  destfile = "fares.zip",
  username = Sys.getenv("NRDP_username"),
  password = Sys.getenv("NRDP_password"),
  url = "https://opendata.nationalrail.co.uk/api/staticfeeds/2.0/fares"
)
}
\arguments{
\item{destfile}{Detestation and name of the zip file}

\item{username}{your username}

\item{password}{your password}

\item{url}{URL of data source}
}
\description{
Downloads fares from https://opendata.nationalrail.co.uk
}
