% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RailDataPortal.R
\name{nrdp_timetable}
\alias{nrdp_timetable}
\title{Download Timetable from National Rail Data Portal}
\usage{
nrdp_timetable(
  destfile = "timetable.zip",
  username = Sys.getenv("NRDP_username"),
  password = Sys.getenv("NRDP_password"),
  url = "https://opendata.nationalrail.co.uk/api/staticfeeds/3.0/timetable"
)
}
\arguments{
\item{destfile}{Detestation and name of the zip file}

\item{username}{your username}

\item{password}{your password}

\item{url}{URL of data source}
}
\description{
Downloads ATOC CIF timetables from https://opendata.nationalrail.co.uk
}
