% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{school_terms}
\alias{school_terms}
\title{Example School term and holiday dates}
\format{
A data frame with 4 columns
}
\usage{
school_terms
}
\description{
Term dates change around the UK and from year to year. This are simply the
Leeds 2022/23 dates repeated for each year. Therefore some inaccuracy will
occur when used historically
}
\keyword{datasets}
