% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atoc_export.R
\name{station2stops}
\alias{station2stops}
\title{Export ATOC stations as GTFS stops.txt}
\usage{
station2stops(station, TI)
}
\arguments{
\item{station}{station SF data frame from the importMSN function}

\item{TI}{TI object from the importMCA function}
}
\description{
Export ATOC stations as GTFS stops.txt
}
\details{
Export ATOC stations as GTFS stops.txt
}
\concept{atoc}
