% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tiplocs}
\alias{tiplocs}
\title{Tiploc Locations}
\format{
A SF data frame for points
}
\usage{
tiplocs
}
\description{
The ATOC data has inaccurate locations for many tiplocs, this is an improved dataset
}
\keyword{datasets}
