% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{UK2GTFS_option_stopProcessingAtUid}
\alias{UK2GTFS_option_stopProcessingAtUid}
\title{UK2GTFS option stopProcessingAtUid}
\usage{
UK2GTFS_option_stopProcessingAtUid(value)
}
\arguments{
\item{value}{option value to be set (char)}
}
\description{
sets/gets a UID value at which processing will stop - used for debugging
}
\details{
If no value passed in will return the current setting of the option. (Usually NULL)
  If value passed in, timetable build processing will stop in atoc_overlay.makeCalendarInner()
  when an exact match for that value is encountered.

  THIS ONLY WORKS WITH ncores==1
}
