% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{UK2GTFS_option_treatDatesAsInt}
\alias{UK2GTFS_option_treatDatesAsInt}
\title{UK2GTFS option treatDatesAsInt}
\usage{
UK2GTFS_option_treatDatesAsInt(value)
}
\arguments{
\item{value}{option value to be set (logical)}
}
\description{
sets/gets a logical value which determines how dates are processed while building calendar - used for debugging
}
\details{
In the critical part of timetable building, handling dates as dates is about half the speed of handling as int
  so we treat them as integers. However that's a complete pain for debugging, so make it configurable.
  if errors are encountered during the timetable build phase, try setting this value to FALSE
}
