% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{UK2GTFS_option_updateCachedDataOnLibaryLoad}
\alias{UK2GTFS_option_updateCachedDataOnLibaryLoad}
\title{UK2GTFS option updateCachedDataOnLibaryLoad}
\usage{
UK2GTFS_option_updateCachedDataOnLibaryLoad(value)
}
\arguments{
\item{value}{option value to be set (logical)}
}
\description{
sets/gets a logical value which determines if the data cached in the library is checked for update when loaded
}
\details{
when child processes are initialised we want to suppress this check, so it is also used for that purpose
}
