% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{update_data}
\alias{update_data}
\title{Update the data inside the UK2GTFS package}
\usage{
update_data()
}
\description{
As UK2GTFS has large datasets that update separately to the R package they
are checked and downloaded at package load time. This function checks for and
downloaded any updated to the data.
}
\details{
Raw data can be viewed and contributed to at
https://github.com/ITSLeeds/UK2GTFS-data
}
