% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_gtfs.R
\name{write_gtfs}
\alias{write_gtfs}
\title{Write GTFS}
\usage{
write_gtfs(gtfs, folder = getwd(), name = "gtfs")
}
\arguments{
\item{gtfs}{named list of data.frames}

\item{folder}{folder to save the gtfs file to}

\item{name}{the name of the zip file, default "gtfs"}
}
\description{
Takes a list of data.frames represneting the GTFS fromat and saves them as GTFS
Zip file.
}
