% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atoc.R
\name{atoc2gtfs}
\alias{atoc2gtfs}
\title{ATOC to GTFS}
\usage{
atoc2gtfs(
  path_in,
  path_out,
  name = "gtfs",
  silent = TRUE,
  ncores = 1,
  locations = tiplocs,
  agency = atoc_agency,
  shapes = FALSE,
  transfers = TRUE
)
}
\arguments{
\item{path_in}{Character, path to ATOC file e.g."C:/input/ttis123.zip"}

\item{path_out}{Character, path to where GTFS files should be saved e.g.
"C:/output"}

\item{name}{Character, name that should be given to the gtfs file, without
the .zip extension (default "gtfs")}

\item{silent}{Logical, should progress messages be surpressed (default TRUE)}

\item{ncores}{Numeric, When parallel processing how many cores to use
(default 1)}

\item{locations}{where to get tiploc locations (see details)}

\item{agency}{where to get agency.txt (see details)}

\item{shapes}{Logical, should shapes.txt be generated (default FALSE)}

\item{transfers}{Logical, should transfers.txt be generated (default TRUE)}
}
\description{
Convert ATOC CIF files to GTFS
}
\details{
Locations

The .msn file contains the physical locations of stations and other TIPLOC
codes (e.g. junctions). However, the quality of the locations is often poor
only accurate to about 1km and occasionally very wrong. Therefore, the
UK2GTFS package contains an internal dataset of the TIPLOC locations with
better location accuracy, which are used by default.

However you can also specify `locations = "file"` to use the TIPLOC locations
in the ATOC data or provide an SF data frame of your own.

Agency

The ATOC files do not contain the necessary information to build the
agency.txt file. Therfore this data is provided with the package. You can
also pass your own data frame of agency information.
}
\concept{main}
