% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atoc.R
\name{atoc2gtfs}
\alias{atoc2gtfs}
\title{ATOC to GTFS}
\usage{
atoc2gtfs(path_in, path_out, name = "gtfs", silent = TRUE,
  ncores = 1, locations = "package")
}
\arguments{
\item{path_in}{Path to ATOC File}

\item{path_out}{Path to where GTFS files should be saved}

\item{name}{name that should be given to the gtfs file, without the .zip extension}

\item{silent}{Logical, should progress be shown}

\item{ncores}{Numeric, When parallel processing how many cores to use}

\item{locations}{character "file" or "package" or path to stops.txt GTFS file, where to get the location data from, default "package"}
}
\description{
ATOC to GTFS
}
\details{
Convert ATOC Files to GTFS

Locations
The .msn file contains the physical locations of stations and other TIPLOC codes (e.g. junctions).
However, the quality of the locations is often poor only accurate to about 1km and occasionally very wrong.
Therefore, the UK2GTFS package contains an internal dataset of the TIPLOC locations with better location accuracy.
}
