% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atoc.R
\name{atoc2gtfs}
\alias{atoc2gtfs}
\title{ATOC to GTFS}
\usage{
atoc2gtfs(
  path_in,
  path_out,
  name = "gtfs",
  silent = TRUE,
  ncores = 1,
  locations = tiplocs,
  agency = atoc_agency,
  shapes = FALSE,
  transferes = TRUE
)
}
\arguments{
\item{path_in}{Path to ATOC File}

\item{path_out}{Path to where GTFS files should be saved}

\item{name}{name that should be given to the gtfs file, without the .zip
extension}

\item{silent}{Logical, should progress be shown}

\item{ncores}{Numeric, When parallel processing how many cores to use}

\item{locations}{deafult tiplocs object inncluded with package, or "file"
or path to stops.txt GTFS file}

\item{agency}{the GTFS agency file, default is taken from the package}

\item{shapes}{Logical, should shapes.txt be generated (default FALSE)}

\item{transfers}{Logical, should transfers.txt be generated (default TRUE)}
}
\description{
ATOC to GTFS
}
\details{
Convert ATOC Files to GTFS

Locations
The .msn file contains the physical locations of stations and other TIPLOC
codes (e.g. junctions). However, the quality of the locations is often poor
only accurate to about 1km and occasionally very wrong Therefore, the
UK2GTFS package contains an internal dataset of the TIPLOC locations with
better location accuracy.

Agency
The ATOC files do not contain the necessary information to build the
agency.txt file. Therfore this data is provided with the package.
}
