% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_naptan.R
\name{get_naptan}
\alias{get_naptan}
\title{Get naptan}
\usage{
get_naptan(
  url = "https://naptan.api.dft.gov.uk/v1/access-nodes?dataFormat=csv",
  naptan_extra = naptan_missing
)
}
\arguments{
\item{url}{character, url to the csv format NaPTAN}

\item{naptan_extra}{data frame of missing stops default uses `naptan_missing`}
}
\value{
data frame of stop locations
}
\description{
Download the NaPTAN stop locations for more information on NaPTAN see
https://data.gov.uk/dataset/ff93ffc1-6656-47d8-9155-85ea0b8f2251/national-public-transport-access-nodes-naptan
}
\details{
TransXchange does not store the location of bus stops, so this
  functions downloads them from the offical DfT source. NaPTAN has some
  missing bus stops which are added by UK2GTFS. See `naptan_missing`
}
