% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs_cleaning.R
\name{gtfs_clean}
\alias{gtfs_clean}
\title{Clean simple errors from GTFS files}
\usage{
gtfs_clean(gtfs, public_only = FALSE)
}
\arguments{
\item{gtfs}{gtfs list}

\item{public_only}{Logical, if TRUE remove routes with route_type missing}
}
\description{
Clean simple errors from GTFS files
}
\details{
Task done:

0. Remove stops with no coordinates
1. Remove stops with no location information
2. Remove trips with less than two stops
3. Remove stops that are never used
4. Replace missing agency names with "MISSINGAGENCY"
5. If service is not public and public_only=TRUE then remove it (freight, 'trips' aka charters)
       (these have a null route_type, so loading into OpenTripPlanner fails if these are present)
6. If public_only=TRUE then remove services with 'train_category' not for public use. e.g. EE (ECS-Empty Coaching Stock)
7. Remove shapes that no longer have any trips
8  Remove frequencies that no longer have any trips
}
