% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs_subset.R
\name{gtfs_clip}
\alias{gtfs_clip}
\title{Clip a GTFS object to a geographical area}
\usage{
gtfs_clip(gtfs, bounds)
}
\arguments{
\item{gtfs}{a gtfs object}

\item{bounds}{an sf data frame of polygons or multipolygons with CRS 4326}
}
\description{
Clips the GTFS file to only include stops within the bounds object, trips
that cross the bouundary of the the object are truncated. Any trips that stop
only once in the bounds are removed completly.
}
