% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs_compress.R
\name{gtfs_compress}
\alias{gtfs_compress}
\title{Reduce file size of a GTFS object}
\usage{
gtfs_compress(gtfs)
}
\arguments{
\item{gtfs}{agtfs objects}
}
\description{
Reduce file size of a GTFS object
}
