% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs_compress.R
\name{gtfs_compress}
\alias{gtfs_compress}
\title{Reduce file size of a GTFS object}
\usage{
gtfs_compress(gtfs)
}
\arguments{
\item{gtfs}{a gtfs object}
}
\value{
a gtfs object
}
\description{
Reduce file size of a GTFS object
}
\details{
by default UK2GTFS trys to preserve id numbers during the conversion
  process to allow back comparions to the original files, e.g.
  `transxchange2gtfs()` retains stop ids from the NAPTAN. However this means
  files sizes are increased. This fucntion replaces ids with intergers and
  thus reduces the size of the gtfs file.
}
