% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs_cleaning.R
\name{gtfs_fast_trips}
\alias{gtfs_fast_trips}
\title{Find fast trips}
\usage{
gtfs_fast_trips(gtfs, maxspeed = 30)
}
\arguments{
\item{gtfs}{list of gtfs tables}

\item{maxspeed}{the maximum allowed speed in metres per second default 30 m/s (about 70 mph)}
}
\description{
Fast trips can idetify problems with the input data or converion process
This fucntion returns trip_ids for trips that exceed max speed.
}
\details{
The fucntion looks a straightline distance between the first and middle stop
to allow for circular routes.
}
