% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs_validate.R
\name{gtfs_force_valid}
\alias{gtfs_force_valid}
\title{Force a GTFS to be valid by removing problems}
\usage{
gtfs_force_valid(gtfs)
}
\arguments{
\item{gtfs}{gtfs object}
}
\description{
Force a GTFS to be valid by removing problems
}
\details{
Actions performed
1. Remove stops with missing location
2. Remove stops from stop_times that are not in stops
3. Remove trips from stop_times that are not in trips
}
