% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs_interpolate_times.R
\name{gtfs_interpolate_times}
\alias{gtfs_interpolate_times}
\title{Interpolate stop times}
\usage{
gtfs_interpolate_times(gtfs, ncores = 1)
}
\arguments{
\item{gtfs}{named list of data.frames}

\item{ncores}{number of cores to use}
}
\description{
Sometimes bus timetables do not give unique stop times to every stop. Instead
several stops in an row are given the same stop time. This function
interpolates the stop times so that each bust stop is given a unique arrival
and departure time.
}
