% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs_merge.R
\name{gtfs_merge}
\alias{gtfs_merge}
\title{merge a list of gtfs files}
\usage{
gtfs_merge(gtfs_list)
}
\arguments{
\item{gtfs_list}{a list of gtfs objects to be merged}
}
\description{
merge a list of gtfs files
}
