% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs_read.R
\name{gtfs_read}
\alias{gtfs_read}
\title{Read GTFS}
\usage{
gtfs_read(path)
}
\arguments{
\item{path}{character, path to GTFS zip folder}
}
\value{
a gtfs object: a named list of data frames, one per GTFS table
}
\description{
Read in a GTFS zip  file
}
