% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs_read.R
\name{gtfs_read}
\alias{gtfs_read}
\title{Read GTFS}
\usage{
gtfs_read(path, stringsAsFactors = FALSE)
}
\arguments{
\item{path}{character, path to GTFS zip folder}

\item{stringsAsFactors}{logical, should character be converted to factors, default FALSE}
}
\description{
Read in a GTFS zip  file
}
