% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs_split.R
\name{gtfs_split}
\alias{gtfs_split}
\title{Split a gtfs file}
\usage{
gtfs_split(gtfs, n_split = 2)
}
\arguments{
\item{gtfs}{a gtfs object}

\item{n_split}{how many to split into default 2}
}
\description{
Split a gtfs file
}
