% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs_split.R
\name{gtfs_split}
\alias{gtfs_split}
\title{Split a GTFS object}
\usage{
gtfs_split(gtfs, n_split = 2)
}
\arguments{
\item{gtfs}{a gtfs object}

\item{n_split}{how many to split into default 2}
}
\value{
a list of GTFS objects
}
\description{
For large regions GTFS fiels can get very big. THis fucntion splits a GTFS
object into a list of GTFS objects. It tries to balance the sizes of the
objects. Splits are made by agency_id so maximumd number of splits equalts
the number of unique agency_ids.
}
