% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs_cleaning.R
\name{gtfs_split}
\alias{gtfs_split}
\title{Split a GTFS object based on trip_ids}
\usage{
gtfs_split(gtfs, trip_ids)
}
\arguments{
\item{gtfs}{gtfs list}

\item{trip_ids}{a vector of trips ids}
}
\description{
Split a GTFS object based on trip_ids
}
