% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stops_per_week_functions.R
\name{gtfs_stop_frequency}
\alias{gtfs_stop_frequency}
\title{Count the number of trips stopping at each stop between two dates}
\usage{
gtfs_stop_frequency(
  gtfs,
  startdate = lubridate::ymd("2020-03-01"),
  enddate = lubridate::ymd("2020-04-30")
)
}
\arguments{
\item{gtfs}{GTFS object from gtfs_read()}

\item{startdate}{Start date}

\item{enddate}{End date}
}
\description{
Count the number of trips stopping at each stop between two dates
}
