% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stops_per_week_functions.R
\name{gtfs_trim_dates}
\alias{gtfs_trim_dates}
\title{Trim a GTFS file between two dates}
\usage{
gtfs_trim_dates(
  gtfs,
  startdate = lubridate::ymd("2020-03-01"),
  enddate = lubridate::ymd("2020-04-30")
)
}
\arguments{
\item{gtfs}{GTFS object from gtfs_read()}

\item{startdate}{Start date}

\item{enddate}{End date}
}
\description{
Trim a GTFS file between two dates
}
