% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stops_per_week_functions.R
\name{gtfs_trips_per_zone}
\alias{gtfs_trips_per_zone}
\title{Trim a GTFS file between two dates}
\usage{
gtfs_trips_per_zone(
  gtfs,
  zone,
  startdate = min(gtfs$calendar$start_date),
  enddate = min(gtfs$calendar$start_date) + 31,
  zone_id = 1,
  by_mode = TRUE,
  ncores = 1,
  time_bands = list(breaks = c(-1, 6, 10, 15, 18, 22, Inf), labels = c("Night",
    "Morning Peak", "Midday", "Afternoon Peak", "Evening", "Night"))
)
}
\arguments{
\item{gtfs}{GTFS object from gtfs_read()}

\item{zone}{SF data frame of polygons}

\item{startdate}{Start date}

\item{enddate}{End date}

\item{zone_id}{Which column in `zone` is the ID column}

\item{by_mode}{logical, disaggregate by mode?}

\item{ncores}{numeric, how many cores to use in parallel processing}

\item{time_bands}{list with two named vectors breaks and labels. Used to
define the time breakdown. Length of breaks must be one greater than length
of labels.}
}
\description{
Trim a GTFS file between two dates
}
