% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stops_per_week_functions.R
\name{gtfs_trips_per_zone}
\alias{gtfs_trips_per_zone}
\title{Trim a GTFS file between two dates}
\usage{
gtfs_trips_per_zone(
  gtfs,
  zone,
  startdate = min(gtfs$calendar$start_date),
  enddate = min(gtfs$calendar$start_date) + 31,
  zone_id = 1,
  by_mode = TRUE,
  ncores = 1
)
}
\arguments{
\item{gtfs}{GTFS object from gtfs_read()}

\item{zone}{SF data frame of polygons}

\item{startdate}{Start date}

\item{enddate}{End date}

\item{zone_id}{Which column in `zone` is the ID column}

\item{by_mode}{logical, disaggregate by mode?}

\item{ncores}{numeric, how many cores to use in parallel processing}
}
\description{
Trim a GTFS file between two dates
}
