% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs_validate.R
\name{gtfs_validate_internal}
\alias{gtfs_validate_internal}
\title{Validate a GTFS object (in R)}
\usage{
gtfs_validate_internal(gtfs)
}
\arguments{
\item{gtfs}{a gtfs object}
}
\description{
Validate a GTFS object (in R)
}
