% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_gtfs.R
\name{gtfs_write}
\alias{gtfs_write}
\title{Write GTFS}
\usage{
gtfs_write(
  gtfs,
  folder = getwd(),
  name = "gtfs",
  stripComma = TRUE,
  stripTab = TRUE,
  stripNewline = TRUE,
  quote = FALSE
)
}
\arguments{
\item{gtfs}{named list of data.frames}

\item{folder}{folder to save the gtfs file to}

\item{name}{the name of the zip file without the .zip extension, default "gtfs"}

\item{stripComma}{logical, should commas be stripped from text, default = TRUE}

\item{stripTab}{logical, should tab be stripped from text, default = TRUE}

\item{stripNewline}{logical, should newline tag be stripped from text, default = TRUE}

\item{quote}{logical, should strings be quoted, default = FALSE}
}
\description{
Takes a list of data frames represneting the GTFS fromat and saves them as GTFS
Zip file.
}
