% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{naptan_missing}
\alias{naptan_missing}
\title{Bus Stop Locations}
\format{
A data frame of 5 columns
}
\usage{
naptan_missing
}
\description{
A database of bus stops that are missing from the NAPTAN but are known to have been used.
}
\keyword{datasets}
