% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NI2gtfs.R
\name{ni2gtfs}
\alias{ni2gtfs}
\title{NI to GTFS}
\usage{
ni2gtfs(path_in, silent = FALSE)
}
\arguments{
\item{path_in}{Path to CIF file}

\item{silent}{Logical, should progress be shown}
}
\value{
A GTFS named list
}
\description{
NI to GTFS
}
\details{
Convert Northern Ireland CIF files to GTFS


}
