% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nptdr.R
\name{nptdr2gtfs}
\alias{nptdr2gtfs}
\title{Import the .CIF file}
\usage{
nptdr2gtfs(
  path = "D:/OneDrive - University of Leeds/Data/UK2GTFS/NPTDR/October-2006.zip",
  silent = FALSE,
  n_files = NULL,
  enhance_stops = TRUE
)
}
\arguments{
\item{path}{Path to zipped folder for NPTDR data}

\item{silent}{Logical, should messages be returned}

\item{n_files}{debug option numerical vector for files to be passed e.g. 1:10}

\item{enhance_stops}{Logical, if TRUE will download current NaPTAN to add in any missing stops}
}
\description{
Import the .CIF file
}
\details{
Imports the CIF file and returns data.frame
}
