% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transxchange2gtfs2.R
\name{transxchange2gtfs}
\alias{transxchange2gtfs}
\title{transxchange2gtfs}
\usage{
transxchange2gtfs(obj, run_debug = TRUE, cal = get_bank_holidays(),
  naptan = get_naptan(), quiet = TRUE)
}
\arguments{
\item{obj}{transxchange object}

\item{run_debug}{logical, should debugs be done?}

\item{cal}{calendar}

\item{naptan}{naptan}

\item{quiet}{logical should messages be displayed}
}
\description{
transxchange2gtfs
}
