% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transxchange.R
\name{transxchange2gtfs}
\alias{transxchange2gtfs}
\title{TransXchange to GTFS}
\usage{
transxchange2gtfs(path_in, path_out, name = "gtfs", silent = TRUE,
  ncores = 1, cal = get_bank_holidays(), naptan = get_naptan())
}
\arguments{
\item{path_in}{Path to ATOC File}

\item{path_out}{Path to where GTFS files should be saved}

\item{name}{name that should be given to the gtfs file, without the .zip extension}

\item{silent}{Logical, should progress be shown}

\item{ncores}{Numeric, When parallel processing how many cores to use}

\item{cal}{Calendar object from get_bank_holidays()}

\item{naptan}{Naptan stop locations from get_naptan()}
}
\description{
TransXchange to GTFS
}
\details{
Convert transxchange files to GTFS


}
