% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transxchange.R
\name{transxchange2gtfs}
\alias{transxchange2gtfs}
\title{TransXchange to GTFS}
\usage{
transxchange2gtfs(
  path_in,
  path_out,
  name = "gtfs",
  silent = TRUE,
  ncores = 1,
  cal = get_bank_holidays(),
  naptan = get_naptan(),
  scotland = "auto"
)
}
\arguments{
\item{path_in}{Path to zipped transxchange files}

\item{path_out}{Path to where GTFS files should be saved}

\item{name}{name that should be given to the gtfs file, without the .zip extension}

\item{silent}{Logical, should progress be shown}

\item{ncores}{Numeric, When parallel processing how many cores to use}

\item{cal}{Calendar object from get_bank_holidays()}

\item{naptan}{Naptan stop locations from get_naptan()}

\item{scotland}{character, should Scottish bank holidays be used?
Can be "auto" (defualt), "yes", "no". If "auto" and path_in ends with "S.zip"
Scottish bank holidays will be used, otherwise England and Wales bank holidays
are used.}
}
\description{
TransXchange to GTFS
}
\details{
Convert transxchange files to GTFS

This is a meta fucntion which aids TransXchange to GTFS conversion. It simple runs
transxchange_import(), transxchange_export(), gtfs_merge(), write_gtfs()

Progress Bars

To minimise overall processing when using mulitple cores the fucntion works
from largest to smallest file.This can mean the progress bar sits a 0% for
quite some time, before starting to move rapidly.
}
