% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transxchange_import6.R
\name{transxchange_import}
\alias{transxchange_import}
\title{Import a TransXchange XML file}
\usage{
transxchange_import(file, run_debug = TRUE, full_import = FALSE)
}
\arguments{
\item{file}{character, path to an XML file e.g. "C:/data/file.xml"}

\item{run_debug}{logical, if TRUE extra checks are performed, default TRUE}

\item{full_import}{logical, if false data no needed for GTFS is excluded}
}
\description{
Import a TransXchange XML file
}
\details{
This function imports the raw transXchange XML files and converts
them to a R readable format.
}
