% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txc_filter_files.R
\name{txc_filter_files}
\alias{txc_filter_files}
\title{Files whose `ServiceCode` cannot be read are always kept.}
\usage{
txc_filter_files(
  files,
  date = Sys.Date(),
  ncores = 1,
  quiet = TRUE,
  resolve_overlaps = TRUE,
  out_dir = NULL
)
}
\description{
Files whose `ServiceCode` cannot be read are always kept.
}
