% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UK2GTFS-package.R
\docType{package}
\name{UK2GTFS-package}
\alias{UK2GTFS}
\alias{UK2GTFS-package}
\title{UK2GTFS: Converts UK transport timetable datasets to GTFS format}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The UK uses a range of odd formats to store timetable data, this package converts them to the nice GTFS format.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/itsleeds/uk2gtfs}
  \item \url{https://itsleeds.git.io/uk2gtfs/}
  \item Report bugs at \url{https://github.com/itsleeds/uk2gtfs/issues}
}

}
\author{
\strong{Maintainer}: Malcolm Morgan \email{m.morgan1@leeds.ac.uk} (\href{https://orcid.org/0000-0002-9488-9183}{ORCID})

Authors:
\itemize{
  \item Owen O'Neill \email{owen@wvr.org.uk} (\href{https://orcid.org/0009-0008-0595-3042}{ORCID})
}

Other contributors:
\itemize{
  \item Adrian Schönig [contributor]
}

}
\keyword{internal}
